# admin/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_admin-template_backend_internal_dto.BatchDeleteReq:
    properties:
      ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - ids
    type: object
...
  github_com_admin-template_backend_internal_dto.CreateMenuReq:
    properties:
      component:
        maxLength: 255
        type: string
      icon:
        maxLength: 64
        type: string
      name:
        maxLength: 64
        type: string
      parent_id:
        type: integer
      path:
        maxLength: 255
        type: string
      perm_code:
        maxLength: 64
        type: string
      sort:
        type: integer
      title:
        maxLength: 64
        type: string
      type:
        enum:
        - 1
        - 2
        - 3
        type: integer
      visible:
        type: boolean
    required:
    - name
    - type
    type: object
...
  github_com_admin-template_backend_internal_dto.MenuResp:
    properties:
      children:
        items:
          $ref: '#/definitions/github_com_admin-template_backend_internal_dto.MenuTree'
        type: array
      component:
        type: string
      created_at:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      perm_code:
        type: string
      sort:
        type: integer
      status:
        type: integer
      title:
        type: string
      type:
        type: integer
      updated_at:
        type: string
      visible:
        type: boolean
    type: object
  github_com_admin-template_backend_internal_dto.MenuTree:
    properties:
      children:
        items:
          $ref: '#/definitions/github_com_admin-template_backend_internal_dto.MenuTree'
        type: array
      component:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      perm_code:
        type: string
      sort:
        type: integer
      title:
        type: string
      type:
        type: integer
      visible:
        type: boolean
    type: object
...
  github_com_admin-template_backend_internal_dto.UpdateMenuReq:
    properties:
      component:
        maxLength: 255
        type: string
      icon:
        maxLength: 64
        type: string
      name:
        maxLength: 64
        type: string
      parent_id:
        type: integer
      path:
        maxLength: 255
        type: string
      perm_code:
        maxLength: 64
        type: string
      sort:
        type: integer
      status:
        enum:
        - 1
        - 2
        type: integer
      title:
        maxLength: 64
        type: string
      type:
        enum:
        - 1
        - 2
        - 3
        type: integer
      visible:
        type: boolean
    type: object
...
  github_com_admin-template_backend_internal_pkg_response.Envelope:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  github_com_admin-template_backend_internal_pkg_response.PageData:
    properties:
      list: {}
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact: {}
  description: RBAC admin backend — auth, users/roles/menus CRUD, operation log, file
    upload, Excel/CSV export.
  title: Admin Template API
  version: "1.0"
paths:
...
  /system/logs:
    get:
      parameters:
      - description: 页码 (默认 1)
        in: query
        name: page
        type: integer
      - description: 每页大小 (默认 20)
        in: query
        name: size
        type: integer
      - description: 模糊搜索用户名/路径/操作
        in: query
        name: keyword
        type: string
      - description: 按 action 模糊过滤
        in: query
        name: action
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.PageData'
              type: object
      security:
      - BearerAuth: []
      summary: 操作日志列表
      tags:
      - System/Logs
  /system/logs/batch-delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID 列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_admin-template_backend_internal_dto.BatchDeleteReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 批量删除日志
      tags:
      - System/Logs
  /system/menus:
    get:
      description: 返回全部菜单的层级结构（不过滤权限），仅用于管理后台配置。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_admin-template_backend_internal_dto.MenuTree'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 菜单树（完整）
      tags:
      - System/Menus
    post:
      consumes:
      - application/json
      parameters:
      - description: 新菜单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_admin-template_backend_internal_dto.CreateMenuReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_admin-template_backend_internal_dto.MenuResp'
              type: object
      security:
      - BearerAuth: []
      summary: 创建菜单
      tags:
      - System/Menus
  /system/menus/{id}:
    delete:
      parameters:
      - description: 菜单 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 删除菜单（递归删除子项）
      tags:
      - System/Menus
    get:
      parameters:
      - description: 菜单 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_admin-template_backend_internal_dto.MenuResp'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 获取单个菜单
      tags:
      - System/Menus
    put:
      consumes:
      - application/json
      parameters:
      - description: 菜单 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 要更新的字段
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_admin-template_backend_internal_dto.UpdateMenuReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_admin-template_backend_internal_dto.MenuResp'
              type: object
      security:
      - BearerAuth: []
      summary: 更新菜单
      tags:
      - System/Menus
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"